Integrate relevant changes into PSE. Check pre initialisation vectors are already handled. Extend a couple of tests to convert from modellist to mapping. TODO : multiscale testing
