+++ Project.toml
@@ -12,6 +12,7 @@ FLoops = "cc61a311-1640-44b5-9fba-1b764f453329"
 Markdown = "d6f4376e-aef5-505a-96c1-9c027394607a"
 MultiScaleTreeGraph = "dd4a991b-8a45-4075-bede-262ee62d5583"
 PlantMeteo = "4630fe09-e0fb-4da5-a846-781cb73437b6"
+SHA = "ea8e919c-243c-51af-8825-aaa63cd721ce"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 Term = "22787eb5-b846-44ae-b979-8e399b8463ab"
@@ -25,6 +26,7 @@ FLoops = "0.2"
 Markdown = "1.9"
 MultiScaleTreeGraph = "0.13, 0.14"
 PlantMeteo = "0.6"
+SHA = "0.7.0"
 Statistics = "1.9"
 Tables = "1"
 Term = "1, 2"
